Generalize replay harness transports for CLI and JSON-RPC (#5)
