+++ apps/server/integration/webAppReplayHarness/index.ts
@@ -1,0 +1,3 @@
+export { createWebAppReplayHarness } from "./createHarness.ts";
+export type { WebAppReplayHarness } from "./createHarness.ts";
+export type { ReplayFixture, ReplayInteraction } from "./types.ts";
